Dessser code gen: add globals (+ envs + params) in environment
Still needs to get the actual values for params etc, though

Also, the way to access variables from the globals module is a bit hackish,
due to the fact that the function module will redefine the same type for
those global vectors. We should instead use Dessser name_type function
to prevent this redefinition, but for this the actual name of that type
is needed, which depends on the backend.
Or, like we define aliases for the variables in that global module, we
also define a type alias for their types.

Ref #1298