Confserver secure protocol: no need to transfer the nonce past the first one
Once the first nonce (aka initial vector) has been transferred to the peer
then the next ones can be incremented on each side.
The encrypted version of the retransmission of the same message will be
distinct due to the nonce and forged retransmission/reordering will fail
to decypher because of unmatched nonce, so everything's fine.