Rework aggr function wrt. list-like operands
Any single operand aggr function can now be given a list and, by default,
it will aggregate the item of that list (in essence behaving not like an
aggregate function) (as opposed to only some of those functions).

The behavior is also more predictable and can be selected via the additional
keyword "immediately" (in addition to "globally" and "locally").

Also, when that's the case, no actual state (neither local nor global) is
actually needed.

Closes #1432