fix: z3 version detection
Previously, the version of z3 was only used for detecting versions of
the form x.y.z where x, y and z are single digit numbers.

However, there are versions of z3 that do not fit this format.

For example, the version 4.8.10 (currently used) is seen as 4.8.1.

Another example: the current last official release is 4.11.2.

This commit fixes this issue.