Upgrade dessser to accept degenerated ClickHouse syntax in CSV
(cherry picked from commit d6a2d3bacb489e437cea74f296c64a47d3af1607)

Unlike the commit that was cherry picked, we still keep csv lines
without the extra spaces after the commas in the example/test.

The reason is that it is still the normal format and we should keep
testing this behavior.